renameing things, switched controller
