+++ src/main/java/frc/robot/subsystems/intake/Intake.java
@@ -5,31 +5,30 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package frc.robot.subsystems;
+package frc.robot.subsystems.intake;
 
 import com.ctre.phoenix.motorcontrol.ControlMode;
 import com.ctre.phoenix.motorcontrol.can.WPI_TalonSRX;
 
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
-import edu.wpi.first.wpilibj.DoubleSolenoid;
-import edu.wpi.first.wpilibj.Joystick;
+// import edu.wpi.first.wpilibj.DoubleSolenoid;
+// import edu.wpi.first.wpilibj.Joystick;
 import edu.wpi.first.wpilibj.Solenoid;
-import edu.wpi.first.wpilibj.TimedRobot;
+// import edu.wpi.first.wpilibj.TimedRobot;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
-import frc.robot.Constants;
-import frc.robot.Robot;
+// import frc.robot.Constants;
+// import frc.robot.Robot;
 import frc.robot.Constants.RobotMap;
 
 import com.ctre.phoenix.motorcontrol.ControlMode;
-import com.ctre.phoenix.motorcontrol.FeedbackDevice;
+// import com.ctre.phoenix.motorcontrol.FeedbackDevice;
 import com.ctre.phoenix.motorcontrol.NeutralMode;
 import com.ctre.phoenix.motorcontrol.SupplyCurrentLimitConfiguration;
 import com.ctre.phoenix.motorcontrol.VelocityMeasPeriod;
-import com.ctre.phoenix.motorcontrol.can.TalonFX;
 
 
 
-public class intake extends SubsystemBase {
+public class Intake extends SubsystemBase {
   /**
    * Creates a new intake.
    */
@@ -39,16 +38,16 @@ public class intake extends SubsystemBase {
   private Solenoid kIntakeSolenoidLeftExtend;
   private Solenoid kIntakeSolenoidLeftRetract;
 
-  private TalonFX m_intakeMotor;
+  private WPI_TalonSRX m_intakeMotor;
   
 
-   public intake() {
-    kIntakeSolenoidRightExtend = new Solenoid(RobotMap.kIntakeSoleniodOne);
-    kIntakeSolenoidRightRetract = new Solenoid(RobotMap.kIntakeSoleniodTwo);
-    kIntakeSolenoidLeftExtend= new Solenoid(RobotMap.kIntakeSoleniodThree);
-    kIntakeSolenoidLeftRetract= new Solenoid(RobotMap.kIntakeSoleniodFour);
+   public Intake() {
+    kIntakeSolenoidRightExtend = new Solenoid(RobotMap.kIntakeSoleniodRightOne);
+    kIntakeSolenoidRightRetract = new Solenoid(RobotMap.kIntakeSoleniodRightTwo);
+    kIntakeSolenoidLeftExtend= new Solenoid(RobotMap.kIntakeSoleniodLeftOne);
+    kIntakeSolenoidLeftRetract= new Solenoid(RobotMap.kIntakeSoleniodLeftTwo);
   
-    m_intakeMotor = new TalonFX(RobotMap.kIntakeTalonFX);
+    m_intakeMotor = new WPI_TalonSRX(RobotMap.kIntakeWPI_TalonSRX);
 
     m_intakeMotor.configFactoryDefault();
 
@@ -59,35 +58,41 @@ public class intake extends SubsystemBase {
     m_intakeMotor.configNeutralDeadband(0.01);
     m_intakeMotor.setNeutralMode(NeutralMode.Coast);
  
-    m_intakeMotor.configSupplyCurrentLimit(new SupplyCurrentLimitConfiguration(true, 45, 80, 0.04));
+    m_intakeMotor.configSupplyCurrentLimit(new SupplyCurrentLimitConfiguration(true, 30, 80, 0.04));
  
-    m_intakeMotor.configSelectedFeedbackSensor(FeedbackDevice.IntegratedSensor);
+    // m_intakeMotor.configSelectedFeedbackSensor(FeedbackDevice.IntegratedSensor);
   }
 
   public void ExtendIntake () {
-    kIntakeSolenoidLeftRetract.set(false);
-    kIntakeSolenoidRightRetract.set(false);
     kIntakeSolenoidRightExtend.set(true);
     kIntakeSolenoidLeftExtend.set(true);
-    
+    kIntakeSolenoidLeftRetract.set(false);
+    kIntakeSolenoidRightRetract.set(false);
+
+    startMotor();
+
   }
 
   public void RetractIntake () {
+
+    stopMotor();
+
     kIntakeSolenoidRightExtend.set(false);
     kIntakeSolenoidLeftExtend.set(false);
     kIntakeSolenoidLeftRetract.set(true);
     kIntakeSolenoidRightRetract.set(true);
   }
 
-  public void CloseIntake () {
-    kIntakeSolenoidLeftRetract.close();
-    kIntakeSolenoidLeftExtend.close();
-    kIntakeSolenoidRightRetract.close();
-    kIntakeSolenoidRightExtend.close();
-  }
-
   public void SetPower (double power) {
     m_intakeMotor.set(ControlMode.PercentOutput, power);
+  }
+
+  public void startMotor () {
+    SetPower(0.5);
+  }
+
+  public void stopMotor () {
+    SetPower(0);
   }
 
   @Override
@@ -95,5 +100,6 @@ public class intake extends SubsystemBase {
     // This method will be called once per scheduler run
     double current = m_intakeMotor.getSupplyCurrent(); 
     SmartDashboard.putNumber("intake current", current);
+
   }
 }
